The Thunderbird manifest was ran through plutil to standardize it.
Note that this removes comments. It appears that most of them came from the manifest template so they can all be removed anyway. As for the others, they appear to be used for marking missing keys which means that they will no be necessary when all keys are supplemented in an upcoming commit.